book-data-pipeline/requirements.txt
apache-airflow==2.7.3
pyspark==3.4.1
pandas==2.0.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
pyarrow==14.0.1
delta-spark==3.0.0
pydantic==2.5.0
great-expectations==0.18.3
jupyter==1.0.0
black==23.11.0
pytest==7.4.3
jellyfish==1.0.3
pydantic-settings==2.1.0